# Speed up DenseSymm by removing dense matvec product