--- a/netket/nn/linear.py
+++ b/netket/nn/linear.py
@@ -219,6 +219,9 @@ def _symmetrizer_col(perms, features):
     case S[ij][kl] == 1. Thus, this method only returns the array of indices `col`
     of shape (n_sites×n_hidden,) satisfying
         W[ij] = w[col[ij]]  <=>  W = w[col].
+
+    See Test/Models/test_nn.py:test_symmetrizer for how this relates to the
+    matrix form of the symmetrizer.
     """
     n_symm, n_sites = perms.shape
     n_hidden = features * n_symm