@PhilipVinc 2021-03-28
I'll try to read carefully what you do tomorrow. 
I also think this index twiddling will be slow on Gpu, but maybe the compiler is super smart.

So this would be a great case for having a decent benchmark of this layer to test...