@femtobit 2021-03-28
@PhilipVinc It doesn't seem too bad on GPU. I did a quick test on my desktop GPU (RTX 3060 Ti using CUDA 11.2). For `N=32` spins, this version is slightly faster (just barely, but given the fear is that this might be slower, that's already a good sign).
```
# N=32, n_samples=1000 on GPU (master)
300/300 [02:01<00:00,  2.47it/s, Energy=-56.692 ± 0.028 [σ²=0.798, R̂=1.0031]]]
# N=32, n_samples=1000 on GPU (this PR)
300/300 [01:58<00:00,  2.54it/s, Energy=-56.692 ± 0.028 [σ²=0.798, R̂=1.0031]]]

# N=32, n_samples=1000 on CPU (master)
300/300 [01:16<00:00,  3.91it/s, Energy=-56.711 ± 0.026 [σ²=0.671, R̂=1.0092]]]
# N=32, n_samples=1000 on CPU (this PR)
300/300 [01:09<00:00,  4.34it/s, Energy=-56.751 ± 0.027 [σ²=0.726, R̂=1.0066]]]
```
I'd love to compare this for a larger system, but storing the full symmetrizer pushes my GPU memory (8 GB) to its limits starting around `N=50`. (Which is also a point in favor of this PR.)

Comparing CPU/GPU just for this branch, I get:
```
# N=70, n_samples=8000 on GPU (this PR)
300/300 [21:37<00:00,  4.33s/it, Energy=-123.847 ± 0.019 [σ²=1.054, R̂=1.0028]]]
# N=70, n_samples=8000 on CPU (this PR)
300/300 [24:05<00:00,  4.82s/it, Energy=-123.878 ± 0.020 [σ²=1.058, R̂=1.0055]]]
```